#[test_span] macro and a couple of test tryouts
