+++ apollo-router-core/src/query_planner/router_bridge_query_planner.rs
@@ -23,7 +23,12 @@ impl RouterBridgeQueryPlanner {
 
 #[async_trait]
 impl QueryPlanner for RouterBridgeQueryPlanner {
-    #[tracing::instrument(name = "plan", level = "debug")]
+    #[tracing::instrument(
+        skip_all,
+        fields(query = %query),
+        name = "plan",
+        level = "debug"
+    )]
     async fn get(
         &self,
         query: String,
