+++ apollo-router-core/src/spec/query.rs
@@ -23,7 +23,14 @@ impl Query {
     ///
     /// This will discard unrequested fields and re-order the output to match the order of the
     /// query.
-    #[tracing::instrument(level = "trace")]
+    #[tracing::instrument(
+        skip_all,
+        fields(
+            data = %response.data.to_string(),
+            errors = %response.errors.iter().map(std::string::ToString::to_string).collect::<Vec<_>>().join(", ")
+        ),
+        level = "trace"
+    )]
     pub fn format_response(&self, response: &mut Response, operation_name: Option<&str>) {
         let data = std::mem::take(&mut response.data);
         match data {
