+++ apollo-router/Cargo.toml
@@ -75,6 +75,7 @@ maplit = "1.0.2"
 mockall = "0.10.2"
 reqwest = { version = "0.11.7", features = ["json", "stream"] }
 test-log = { version = "0.2.8", default-features = false, features = ["trace"] }
+test-span = { path = "../test-span" }
 tracing-subscriber = { version = "0.3", default-features = false, features = [
     "env-filter",
     "fmt",
