+++ test-span-macro/Cargo.toml
@@ -1,0 +1,12 @@
+[package]
+name = "test-span-macro"
+version = "0.1.0"
+edition = "2021"
+
+[lib]
+proc-macro = true
+
+[dependencies]
+proc-macro2 = "1.0"
+quote = { version = "1.0" }
+syn = { version = "1.0", features = ["full"] }
