+++ test-span/Cargo.toml
@@ -1,0 +1,23 @@
+[package]
+name = "test-span"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+daggy = { version = "0.8.0", features = ["serde-1", "stable_dag"] }
+serde = { version = "1.0.131", features = ["derive"] }
+serde_json = "1.0.72"
+tokio = { version = "1.14.0", features = ["full"] }
+tracing = "0.1.29"
+tracing-futures = "0.2.5"
+tracing-subscriber = "0.3.3"
+test-span-macro = { path = "../test-span-macro" }
+
+[dev-dependencies]
+insta = "1.8.0"
+
+[[test]]
+name = "tests"
+path = "tests/tests.rs"
